+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - Independent recovery analysis [#5](https://github.com/IN-CORE/pyincore-incubator/issues/5)
+- Electric power network recovery analysis [#10](https://github.com/IN-CORE/pyincore-incubator/issues/10)
 - GitHub action to check linting [#16](https://github.com/IN-CORE/pyincore-incubator/issues/16)
 
 ### Fixed